Devops: Add unit tests for calculation and parser plugin
Unit tests are added for the `CalcJob` and `Parser` plugins. The
reference data for the parser tests were taken from the data of the
AiiDA archive that was part of the AiiDA common workflow paper:
https://archive.materialscloud.org/record/file?filename=acwf-verification_unaries-verification-PBE-v1_results_wien2k.aiida&record_id=1770

The `pytest` package is used for the tests with `pytest-regressions`
being used to compare results with reference files.