Devops: Manually fix parsers reference files to pass tests
The reference files for the parser tests were taken from calculations
that were run with WIEN2k v21, however, the parser plugin had already
been updated to be compatible with WIEN2k v23. One of the most important
differences is that WIEN2k v21 writes the convergence line as:

    ec cc and fc_conv 1 1 1

In WIEN2k v23, stress convergence was added to this line:

    ec cc fc and str_conv 1 1 1 1

Ideally, the reference output files are regenerated with WIEN2k v23 but
I have no access to this now. Instead, the reference output files are
manually corrected to match the v23 format.

The `test_failed_warning_other` test also failed because the only
warning it had mapped to the `WARNING_QTL_B` error code. This warning
was manually updated with a made up warning to pass the test.